extractProcess/requirements.txt
# Function dependencies, for example:
# package>=version
wikipedia==1.4.0 
google-api-python-client==1.7.8
flask==1.0.2
beautifulsoup4==4.7.1 
slackclient==2.0.1